requirements.txt
google-adk
gymnasium[box2d]
python-dotenv
imageio
numpy
